chore: refactor main and startup modules
